refactor(confirmations): consolidate External-app detection into one hook
The three "Request from" surfaces had each grown their own copy of the
unverifiable-origin logic, and they had already diverged: OriginRow
checked only isExternalAppOrigin — and against useApprovalInfo's url,
which for signature requests is the self-reported page URL
(requestData.meta.url), never a UUID — so its signature branches could
never classify an SDK / MWP request as external; NetworkAndOriginRow
checked origin + request_source; InfoSectionOriginAndDetails duplicated
that inline.

Add a useIsExternalAppRequest hook that derives the answer from the
actual request origin (signature / transaction / transaction batch) plus
the transport request_source, and use it in all three rows. Adding a new
unverifiable transport is now a one-place change, closing the latent
fail-open divergence flagged in review.
