+++ app/components/Views/confirmations/components/info/typed-sign-v3v4/info-section-origin-and-details/info-section-origin-and-details.tsx
@@ -15,10 +15,7 @@ import {
   isRecognizedPermit,
   parseAndNormalizeSignTypedDataFromSignatureRequest,
 } from '../../../../utils/signature';
-import {
-  isExternalAppOrigin,
-  isExternalAppRequestSource,
-} from '../../../../utils/origin';
+import { useIsExternalAppRequest } from '../../../../hooks/useIsExternalAppRequest';
 import { useSignatureRequest } from '../../../../hooks/signatures/useSignatureRequest';
 import useApprovalRequest from '../../../../hooks/useApprovalRequest';
 import { View } from 'react-native';
@@ -40,24 +37,10 @@ export const InfoSectionOriginAndDetails = () => {
   const signatureRequest = useSignatureRequest();
   const isPermit = isRecognizedPermit(signatureRequest);
 
-  // `request_source` is the transport the request arrived on. It is only
-  // populated on signature requests; `meta` is not on the persisted
-  // MessageParams union type, hence the structural cast.
-  const requestSource = (
-    signatureRequest?.messageParams as
-      | { meta?: { analytics?: { request_source?: string } } }
-      | undefined
-  )?.meta?.analytics?.request_source;
-
   // For requests where we cannot verify the dapp's identity, display a generic
   // "External app" label rather than the raw origin — same handling as
-  // OriginRow / NetworkAndOriginRow. This covers `ethereum:` deeplinks /
-  // scanned QR codes, MetaMask SDK and MetaMask Connect (MWP) connections
-  // (origin is a bare connection UUID), and any remote transport whose origin
-  // is self-reported and therefore unverifiable (SDK v1, MWP, WalletConnect)
-  // as identified by `request_source`.
-  const isExternalApp =
-    isExternalAppOrigin(origin) || isExternalAppRequestSource(requestSource);
+  // OriginRow / NetworkAndOriginRow.
+  const isExternalApp = useIsExternalAppRequest();
 
   const parsedData =
     parseAndNormalizeSignTypedDataFromSignatureRequest(signatureRequest);
