+++ app/components/Views/confirmations/components/rows/transactions/network-and-origin-row/network-and-origin-row.tsx
@@ -11,11 +11,8 @@ import Text, {
 import { useStyles } from '../../../../../../../component-library/hooks';
 import { strings } from '../../../../../../../../locales/i18n';
 import { useApprovalInfo } from '../../../../hooks/useApprovalInfo';
+import { useIsExternalAppRequest } from '../../../../hooks/useIsExternalAppRequest';
 import { MMM_ORIGIN } from '../../../../constants/confirmations';
-import {
-  isExternalAppOrigin,
-  isExternalAppRequestSource,
-} from '../../../../utils/origin';
 import InfoSection from '../../../UI/info-row/info-section';
 import InfoRow from '../../../UI/info-row/info-row';
 import Address from '../../../UI/info-row/info-value/address';
@@ -34,13 +31,13 @@ export const NetworkAndOriginRow = () => {
   const origin =
     transactionMetadata?.origin || signatureRequest?.messageParams?.origin;
 
-  // `request_source` is the transport the request arrived on. It is only
-  // populated on signature requests; transactions persist only `origin`.
-  const requestSource = (
-    signatureRequest?.messageParams as
-      | { meta?: { analytics?: { request_source?: string } } }
-      | undefined
-  )?.meta?.analytics?.request_source;
+  // For requests where we cannot verify the dapp's identity, display a generic
+  // "External app" label rather than the raw origin. This covers `ethereum:`
+  // deeplinks / scanned QR codes (origin === 'deeplink' / 'qr-code'), MetaMask
+  // SDK and MetaMask Connect (MWP) connections (origin is a bare connection
+  // UUID), and any remote transport whose origin is self-reported and therefore
+  // unverifiable (SDK v1, MWP, WalletConnect) as identified by `request_source`.
+  const isExternalApp = useIsExternalAppRequest();
 
   const isDappOrigin = origin !== MMM_ORIGIN;
 
@@ -48,14 +45,6 @@ export const NetworkAndOriginRow = () => {
     return null;
   }
 
-  // For requests where we cannot verify the dapp's identity, display a generic
-  // "External app" label rather than the raw origin. This covers `ethereum:`
-  // deeplinks / scanned QR codes (origin === 'deeplink' / 'qr-code'), MetaMask
-  // SDK and MetaMask Connect (MWP) connections (origin is a bare connection
-  // UUID), and any remote transport whose origin is self-reported and therefore
-  // unverifiable (SDK v1, MWP, WalletConnect) as identified by `request_source`.
-  const isExternalApp =
-    isExternalAppOrigin(origin) || isExternalAppRequestSource(requestSource);
   const displayedOrigin = isExternalApp
     ? strings('confirm.label.external_app')
     : origin;
